backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
python-dotenv==1.0.0
requests==2.28.2
gunicorn==20.1.0  # For production deployment 